backup and hash function to prevent lost data.
this update is for fix issue in my code and prevent lost data. 

I made an hash function to see if the bdd was modify. 
If so, the programme look into his backup to uncript the bdd with the last key.
else, the programe uncript the bdd, and he update his backup with his this key.