renaming keylogger file
i juste rename this file because first of all, this is not really a keylogger, because i don't save the key pres. And secondly, my anti virus juste delated my file. So i rename it in "selection" because it discribe more want it is really and my anti virus dosn't dellated it.
(so this is not a majeur improuvement)
sorry for the time that you lose while reading this lines... =)