the exe feature
i juste add the execution of application by a shortcut. But, i don't do the selection.
In fact, just have the choice of using the program in password gestionnaire mode or in execution shortcut software.
but of curse it will change soon in the future.
i just need to remake the selection modelle to andelle with the différente features and "mode" this software can have.